pyproject.toml
[project]
name = "disc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.10.7",
    "discord-py>=2.4.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
